requirements.txt
streamlit
pandas
plotly
databricks-sdk
requests
geopy